pyproject.toml
[tool.poetry]
name = "car-recognition"
version = "0.1.0"
description = ""
authors = ["Stanislav <sta-tun-slav@yandex.ru>"]
readme = "README.md"
packages = [{include = "src"}]


[tool.poetry.scripts]
start = "src.main:run"
set_tables = "src.db:create_tables"

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12.0"
opencv-python = "^4.8.1.78"
scikit-image = "^0.22.0"
matplotlib = "^3.8.1"
tensorflow = "^2.14.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pyodbc = "^5.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
